Rename joinAddr and move FromDomainProject into pkg/util
