+++ pkg/util/util.go
@@ -156,3 +156,15 @@ func StringTRUE(s string) bool {
 	}
 	return false
 }
+
+func FromDomainProject(domainProject string) (domain, project string) {
+	if i := strings.Index(domainProject, "/"); i >= 0 {
+		return domainProject[:i], domainProject[i+1:]
+	}
+	return domainProject, ""
+}
+
+func ToDomainProject(domain, project string) (domainProject string) {
+	domainProject = domain + "/" + project
+	return
+}
