+++ scctl/pkg/plugin/get/instance/printer.go
@@ -19,7 +19,6 @@ import (
 	"github.com/apache/servicecomb-service-center/pkg/util"
 	"github.com/apache/servicecomb-service-center/scctl/pkg/model"
 	"github.com/apache/servicecomb-service-center/scctl/pkg/writer"
-	"github.com/apache/servicecomb-service-center/server/core"
 	"time"
 )
 
@@ -61,7 +60,7 @@ func (s *InstanceRecord) AgeString() string {
 }
 
 func (s *InstanceRecord) Domain() string {
-	domain, _ := core.FromDomainProject(s.DomainProject)
+	domain, _ := util.FromDomainProject(s.DomainProject)
 	return domain
 }
